Change Viewer to StructureViewer; add docstrings
